Safari-611.1.10/LayoutTests/accessibility/svg-shape-labelled-expected.txt


This tests that SVG shapes are exposed to ax tree as images if there are labels describing them.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


circle1: AXRole: AXImage
circle2: AXRole: AXImage
circle3: Element not exposed
PASS !image1 is true
PASS image2.role is 'AXRole: AXImage'
PASS image3.role is 'AXRole: AXImage'
PASS successfullyParsed is true

TEST COMPLETE
